refactor: retryUntilValid 로직을 Input View로 이동
